#1749 Correct code smell
Delete copy operator and constructor
Define move operator and constructor

+++ dynawo/sources/Modeler/DataInterface/PowSyblIIDM/DYNLineInterfaceIIDM.cpp
@@ -53,7 +53,9 @@ LineInterfaceIIDM::LineInterfaceIIDM(powsybl::iidm::Line& line) : lineIIDM_(line
 }
 
 LineInterfaceIIDM::~LineInterfaceIIDM() {
-  destroyActiveSeasonExtension_(activeSeasonExtension_);
+  if (destroyActiveSeasonExtension_) {
+    destroyActiveSeasonExtension_(activeSeasonExtension_);
+  }
 }
 
 void
